refactor: improve code organization and architecture
