+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fireworks-pixi-project",
-  "version": "0.1.0",
+  "version": "0.2.1",
   "private": true,
   "type": "module",
   "scripts": {
