--- src/utils/math.ts
@@ -1,3 +1,0 @@
-export function getRandomInt(min: number, max: number): number {
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-}
